remote/requirements_gpu.txt
absl-py==2.0.0
accelerate==0.23.0
albumentations==1.3.1
attrs==23.1.0
beautifulsoup4==4.12.2
blurhash==1.1.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cmake==3.27.6
coloredlogs==15.0.1
contourpy==1.1.1
controlnet-aux==0.0.7
cycler==0.12.0
Cython==3.0.3
decorator==5.1.1
peft @ git+https://github.com/huggingface/peft.git
diffusers @ git+https://github.com/huggingface/diffusers.git@84e5cc596c67afd2863131f5fa309ebaa994624d
easydict==1.10
einops==0.7.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.43.0
fsspec==2023.9.2
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
imageio==2.31.4
importlib-metadata==6.8.0
insightface==0.7.3
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==17.0.1
MarkupSafe==2.1.3
Mastodon.py==1.8.1
matplotlib==3.8.0
mediapipe==0.10.5
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.14.1
onnxruntime-gpu==1.16.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.1
prettytable==3.9.0
protobuf==3.20.3
psutil==5.9.5
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
python-magic==0.4.27
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.6.12
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.3
triton==2.0.0
typing_extensions==4.8.0
urllib3==2.0.7
xformers==0.0.22
zipp==3.17.0
